refactor(Internal): moved some stuff into new modules
