+++ src/Text/Gigaparsec/Internal/Errors/DefuncTypes.hs
@@ -1,22 +1,21 @@
 {-# LANGUAGE Trustworthy #-}
-{-# LANGUAGE BangPatterns, CPP, DataKinds, GADTs, UnboxedSums #-}
+{-# LANGUAGE BangPatterns, CPP, DataKinds, GADTs #-}
 {-# OPTIONS_GHC -Wno-partial-fields -Wno-all-missed-specialisations -Wno-missing-import-lists #-}
 {-# OPTIONS_HADDOCK hide #-}
 -- Yes, this is redundant, however, it is necessary to get the UNPACK to fire
 {-# OPTIONS_GHC -Wno-redundant-strictness-flags #-}
 #include "portable-unlifted.h"
-module Text.Gigaparsec.Internal.Errors.DefuncError (
-    module Text.Gigaparsec.Internal.Errors.DefuncError
+module Text.Gigaparsec.Internal.Errors.DefuncTypes (
+    module Text.Gigaparsec.Internal.Errors.DefuncTypes
   ) where
+
+import Text.Gigaparsec.Internal.Errors.CaretControl (CaretWidth, Span)
+import Text.Gigaparsec.Internal.Errors.ErrorItem (ExpectItem)
 
 import Data.Word (Word32)
 import Data.Set (Set)
 
 CPP_import_PortableUnlifted
-
--- TODO: remove
-type Span :: *
-type Span = Word
 
 type ErrKind :: *
 data ErrKind = Vanilla | Specialised
@@ -43,9 +42,9 @@ data DefuncError_ k where
 type BaseError :: ErrKind -> UnliftedDatatype
 data BaseError k where
   ClassicSpecialised :: ![String] -> BaseError 'Specialised
-  Expected :: {-!(Set ExpectItem) ->-} {-# UNPACK #-} !Span -> BaseError 'Vanilla
-  ExpectedWithReason :: {-!(Set ExpectItem) ->-} !String -> {-# UNPACK #-} !Span -> BaseError 'Vanilla
-  Unexpected :: {-!(Set ExpectItem) -> !String -> {-# UNPACK #-} !CaretWidth -> -} BaseError 'Vanilla
+  Expected :: !(Set ExpectItem) -> {-# UNPACK #-} !Span -> BaseError 'Vanilla
+  ExpectedWithReason :: !(Set ExpectItem) -> !String -> {-# UNPACK #-} !Span -> BaseError 'Vanilla
+  Unexpected :: !(Set ExpectItem) -> !String -> {-# UNPACK #-} !CaretWidth -> BaseError 'Vanilla
   Empty :: {-# UNPACK #-} !Span -> BaseError 'Vanilla
   EmptyWithReason :: !String -> {-# UNPACK #-} !Span -> BaseError 'Vanilla
 
@@ -55,7 +54,7 @@ data ErrorOp k where
   AdjustCaret :: !(DefuncError 'Specialised)
               -> !(DefuncError 'Vanilla) -- ^ caretAdjuster
               -> ErrorOp 'Specialised
-  WithHints   :: !(DefuncError 'Vanilla) {- -> !DefuncHints -}-> ErrorOp 'Vanilla
+  WithHints   :: !(DefuncError 'Vanilla) -> !DefuncHints -> ErrorOp 'Vanilla
   WithReason  :: !(DefuncError 'Vanilla) -> !String -> ErrorOp 'Vanilla
   WithLabel   :: !(DefuncError 'Vanilla) -> !(Set String) -> ErrorOp 'Vanilla
   Amended     :: {-# UNPACK #-} !Word -- ^ line
@@ -66,3 +65,10 @@ data ErrorOp k where
   Dislodged   :: {-# UNPACK #-} !Word -- ^ by
               -> !(DefuncError k) -> ErrorOp k
   Lexical     :: !(DefuncError 'Vanilla) -> ErrorOp 'Vanilla
+
+type DefuncHints :: UnliftedDatatype
+data DefuncHints where
+  Blank :: DefuncHints
+  Replace :: !(Set String) -> DefuncHints
+  Merge :: !DefuncHints -> !DefuncHints -> DefuncHints
+  AddErr :: !DefuncHints -> !(DefuncError 'Vanilla) -> DefuncHints
